Move the form into its own container
